move consensus tests to under tests folder
